% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{adicionar_passo_previsao}
\alias{adicionar_passo_previsao}
\title{Adiciona Passo De Previsao}
\usage{
adicionar_passo_previsao(dt_prev)
}
\arguments{
\item{dt_prev}{´data.table´ contendo os dados das variaveis previstas do modelo NWP}
}
\value{
´data.table´ das usinas simuladas e as respectivas variaveis previstas do modelo NWP
}
\description{
Obtem as previsoes dos modelos NWP para as usinas simuladas. Para isso, realiza a associacao das coordenadas das
usinas com quadricula correspondente do modelo NWP.
E escolhida a quadricula que apresenta menor distancia euclidiana entre seu centroide e a coordenada da usina.
}
