% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{ajusta_arima}
\alias{ajusta_arima}
\title{Ajusta modelo ARIMA simples}
\usage{
ajusta_arima(dt, nlmod0)
}
\arguments{
\item{dt}{data.table com variavel \code{ger_obs_norm}.}

\item{nlmod0}{modelo dummy utilizado em caso de erro.}
}
\value{
modelo ajustado do tipo \code{Arima}.
}
\description{
Ajusta modelo ARIMA simples
}
