% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{ajusta_arimax}
\alias{ajusta_arimax}
\title{Ajusta modelo ARIMAX com variaveis exogenas}
\usage{
ajusta_arimax(dt, nlmod0)
}
\arguments{
\item{dt}{data.table com variavel \code{ger_obs_norm}, \code{irrad_prev_norm},
\code{temp_prev_norm} e \code{umid_prev_norm}.}

\item{nlmod0}{modelo dummy utilizado em caso de erro.}
}
\value{
modelo ajustado do tipo \code{Arima}.
}
\description{
Ajusta modelo ARIMAX com variaveis exogenas
}
