% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{ajusta_modelo_fallback}
\alias{ajusta_modelo_fallback}
\title{Ajusta Modelo Fallback (Zeros)}
\usage{
ajusta_modelo_fallback(dt_treino_filt)
}
\arguments{
\item{dt_treino_filt}{data.table com dados filtrados}
}
\value{
Objeto lm com coeficientes zerados
}
\description{
Cria modelo de regressao linear neutro para uso como fallback.
}
\keyword{internal}
