% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{ajusta_rls_rlm}
\alias{ajusta_rls_rlm}
\title{Ajusta Modelos RLS e RLM}
\usage{
ajusta_rls_rlm(dt_treino_filt, nlmod0)
}
\arguments{
\item{dt_treino_filt}{data.table com dados filtrados}

\item{nlmod0}{modelo fallback}
}
\value{
Lista com nlmod1 e nlmod2, ou NULL se dados insuficientes
}
\description{
Tenta ajustar regressao linear simples (RLS) e multipla (RLM).
Retorna NULL se dados insuficientes.
}
\keyword{internal}
