% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{associa_nwp_usina}
\alias{associa_nwp_usina}
\title{Associa Usinas as Quadriculas NWP}
\usage{
associa_nwp_usina(dt_prev, dt_usinas)
}
\arguments{
\item{dt_prev}{data.table com previsoes NWP contendo:
\describe{
\item{id_modelo_nwp}{Identificador do modelo NWP}
\item{latitude}{Latitude do ponto de grade}
\item{longitude}{Longitude do ponto de grade}
\item{data_hora_rodada}{Data/hora da rodada}
\item{data_hora_previsao}{Data/hora da previsao}
\item{valor}{Valor previsto}
}}

\item{dt_usinas}{data.table com cadastro de usinas contendo:
\describe{
\item{id_usina}{Identificador da usina}
\item{latitude}{Latitude da usina}
\item{longitude}{Longitude da usina}
}}
}
\value{
data.table com previsoes associadas as usinas, incluindo
coluna \code{id_usina}
}
\description{
Mapeia cada usina para a quadricula mais proxima do modelo NWP,
usando distancia euclidiana entre coordenadas.
}
\details{
Para cada usina, a funcao:
\enumerate{
\item Calcula distancia euclidiana para todos os pontos de grade
\item Seleciona o ponto mais proximo por modelo NWP
\item Atribui previsoes desse ponto a usina
}

A coluna \code{id_usina} e posicionada como segunda coluna no resultado.
}
\keyword{internal}
