% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{calcula_erros}
\alias{calcula_erros}
\title{Calcula erros medios in-sample dos modelos}
\usage{
calcula_erros(dt, nlmod1, nlmod2)
}
\arguments{
\item{dt}{data.table com variavel \code{ger_obs_norm}, \code{irrad_prev_norm},
\code{temp_prev_norm} e \code{umid_prev_norm}..}

\item{nlmod1}{Modelo ARIMA.}

\item{nlmod2}{Modelo ARIMAX.}
}
\value{
Lista com erro medio de cada modelo.
}
\description{
Calcula erros medios in-sample dos modelos
}
