% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{carrega_modelo_rds}
\alias{carrega_modelo_rds}
\title{Carrega Artefato de Modelo}
\usage{
carrega_modelo_rds(modelo_previsao, id_usina, diretorio)
}
\arguments{
\item{modelo_previsao}{Tipo do modelo ("arimax", "fisico_estimado")}

\item{id_usina}{Identificador da usina}

\item{diretorio}{Caminho para diretorio de artefatos}
}
\value{
Objeto do modelo carregado
}
\description{
Le arquivo RDS contendo modelo treinado para uma usina.
}
\keyword{internal}
