% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{clear_log_context}
\alias{clear_log_context}
\title{Remove Contexto Estruturado do Logger}
\usage{
clear_log_context()
}
\value{
\code{invisible(NULL)}
}
\description{
Remove o filtro \code{"pfvsh_ctx"} do logger. No-op se o filtro nao existir.
}
\seealso{
\code{\link[=set_log_context]{set_log_context()}}
}
