% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{completa_datas}
\alias{completa_datas}
\title{Completa Datas Da Previsao Final}
\usage{
completa_datas(dt, discretizacao)
}
\arguments{
\item{discretizacao}{Intervalo de tempo para a sequencia (e.g., "30 min", "1 hour")}
}
\value{
data.table com colunas de \code{data_hora_previsao}
preenchida com sequencia completa
}
\description{
Gera data.table com sequencia completa de timestamps da previsao final
}
\keyword{internal}
