% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{config_names}
\alias{config_names}
\title{Retorna Nomes das Chaves Obrigatorias}
\usage{
config_names()
}
\value{
Vetor de strings com nomes das chaves de configuracao
}
\description{
Retorna Nomes das Chaves Obrigatorias
}
\keyword{internal}
