% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{config_types}
\alias{config_types}
\title{Retorna Tipos Esperados por Chave}
\usage{
config_types()
}
\value{
Lista nomeada com tipos esperados para cada chave
}
\description{
Retorna Tipos Esperados por Chave
}
\keyword{internal}
