% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{cria_dt_auxiliar}
\alias{cria_dt_auxiliar}
\title{Cria Data.table Auxiliar}
\usage{
cria_dt_auxiliar(dif, dt_completo)
}
\arguments{
\item{dif}{´data.table´ contendo a data ausente nos dados de previsao numerica para a qual se deseja
criar o data.table de previsoes NA. Contem tambem as demais informacoes necessaria para criar o}

\item{dt_completo}{´data.table´ completo com as previsoes meteorologicas para referencia}
}
\value{
´data.table´ nor formato adequado para inclusao nos dados de previsao meteorologica a ser usado
}
\description{
Auxiliar da funcao ´preenche_ausencia_previsao´. Cria um ´data.table´ auxiliar
com as datas ausentes nos dados originais de previsao meteorologica. As previsoes meteorologicas
para essas datas sao preenchidas com ´NA´.
}
