% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cria_sequencia_datas}
\alias{cria_sequencia_datas}
\title{Cria Sequencia De Datas}
\usage{
cria_sequencia_datas(dt, discretizacao)
}
\arguments{
\item{dt}{data.table com dados de previsao}

\item{discretizacao}{intervalo de tempo entre datas, ex: "30 min"}
}
\value{
data.table com sequencia completa de datas
}
\description{
Identifica as datas de inicio e fim das series temporais
e cria dt com a sequencia completa de datas
}
