% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{define_hor_prev}
\alias{define_hor_prev}
\title{Gera vetor de datas-alvo de previsao}
\usage{
define_hor_prev(data_referencia, v_horizonte)
}
\arguments{
\item{data_referencia}{\code{Date} ou \code{character} coerente com formato de data.
Representa a data-base da execucao da previsao (ex: "2025-11-07").}

\item{v_horizonte}{Vetor indicando os horizontes de previsao.}
}
\value{
Vetor de classe \code{Date} contendo as datas alvo de previsao.
}
\description{
Esta funcao cria um vetor de datas correspondentes aos horizontes de previsao,
com base em uma data de referencia (default: o dia de execucao da previsao)
e em um vetor de horizontes (D+0 a D+9).
}
\examples{
define_hor_prev("2025-11-07", c("D+0", "D+1"))
# Retorna: 2025-11-07, 2025-11-08

}
