% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{desnormaliza_variaveis}
\alias{desnormaliza_variaveis}
\title{Desnormaliza variaveis com base em estatisticas de normalizacao}
\usage{
desnormaliza_variaveis(dt, stats_norm)
}
\arguments{
\item{dt}{data.table com colunas normalizadas (ex: *_norm)}

\item{stats_norm}{data.table com colunas: variavel, med, sd}
}
\value{
data.table com colunas desnormalizadas (sem sufixo "_norm")
}
\description{
Desnormaliza variaveis com base em estatisticas de normalizacao
}
