% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{elimina_dados_invalidos}
\alias{elimina_dados_invalidos}
\title{Elimina Dados Invalidos}
\usage{
elimina_dados_invalidos(dt)
}
\arguments{
\item{dt}{'data.table' contendo os dados com as series temporais}
}
\value{
'data.table' contendo os dados com as series temporais apos eliminacao de dados invalidos
}
\description{
Identifica dados invalidos nas series temporais e elimina da amostra.
A funcao avalia os valores das series temporais das variaveis contidas no data.table,
identifica posicoes com valores 999 ou NA e elimina de todas as variaveis concomitantemente.
}
