% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{extract_strategy_name}
\alias{extract_strategy_name}
\title{Extrai Nome da Estrategia do Plano Atual}
\usage{
extract_strategy_name(plan_obj)
}
\arguments{
\item{plan_obj}{objeto de plano retornado por \code{future::plan()}}
}
\value{
character com o nome da estrategia
}
\description{
Extrai Nome da Estrategia do Plano Atual
}
\keyword{internal}
