% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gera_combinacoes_modelo}
\alias{gera_combinacoes_modelo}
\title{Gera combinacoes entre nwp, passo de previsao e meia-hora}
\usage{
gera_combinacoes_modelo(
  v_modelos_nwp,
  v_horizonte,
  periodo_ger,
  v_modelos_previsao
)
}
\arguments{
\item{v_modelos_nwp}{vetor com os nomes dos modelos meteorologicos, ex: c("GFS")}

\item{v_horizonte}{vetor de caracteres com os passos de previsao, ex: c("D+0", "D+1")}

\item{periodo_ger}{vetor de horarios no formato HH:MM, ex: c("12:00", "12:30")}

\item{v_modelos_previsao}{vetor com os modelos de previsao}
}
\value{
lista com as colunas:
\itemize{
\item id_modelo_nwp - modelo meteorologico
\item horiz_prev - passo de previsao
\item periodo_ger - meias-horas com geracao solar
}
}
\description{
Cria uma lista com todas as combinacoes possiveis entre modelos meteorologicos,
passos de previsao e meias-horas
}
\examples{
\dontrun{
v_nwp <- c("GFS", "ECMWF")
v_horiz <- c("D+0", "D+1")
v_hor_ger <- c("12:00", "12:30")
gera_combinacoes_modelo(v_nwp, v_horiz, v_hor_ger, "arimax")
}
}
