% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{get_dataset}
\alias{get_dataset}
\title{Cria Dataset para Treinamento}
\usage{
get_dataset(args, conn)
}
\arguments{
\item{args}{Lista com \code{ids_usinas} e \code{modelos_NWP}}

\item{conn}{Conexao com banco de dados}
}
\value{
Lista com \code{ger_obs} e \code{irrad_prev}
}
\description{
Carrega dados para treinamento dos modelos.
}
\keyword{internal}
