% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{identifica_periodo_ger}
\alias{identifica_periodo_ger}
\title{Identifica periodo do dia com geracao solar}
\usage{
identifica_periodo_ger(dad_usina, ger_usi, fator_tol_ger, fator_tol_horas)
}
\arguments{
\item{dad_usina}{data.table contendo pelo menos:
\itemize{
\item id_usina
\item potencia instalada
}}

\item{ger_usi}{data.table contendo pelo menos:
\itemize{
\item data_hora_observacao
\item valor
}}

\item{fator_tol_ger}{fator minimo da capacidade instalada da usina para considerar
que houve geracao}

\item{fator_tol_horas}{fator minimo dos dias analisados com geracao}
}
\value{
vetor de characteres com as meias-horas ("HH:MM")
}
\description{
A funcao identifica horarios ("HH:MM") em que ha geracao
ao longo de um ano.
Util para definir periodo em que havera treinamento de modelos
e previsao de geracao solar
}
\details{
A funcao segue as etapas:
\itemize{
\item extrai a meia-hora, hora_min, de cada observacao
\item seleciona os dias para analise
\item para cada meia-hora, conta o numero de dias com geracao superior ao limiar estabelecido
\item retorna as meias-horas com geracao em pelo menos \code{fator_tol_dias*janela_dias} dias.
}
}
