% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.r
\name{is_plant_error}
\alias{is_plant_error}
\title{Verifica se Objeto e um Erro de Usina}
\usage{
is_plant_error(x)
}
\arguments{
\item{x}{objeto a verificar}
}
\value{
\code{TRUE} se \code{x} tiver classe \code{"plant_error"}, \code{FALSE} caso contrario
}
\description{
Verifica se Objeto e um Erro de Usina
}
