% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{is_relative_path}
\alias{is_relative_path}
\title{Verifica se Caminho e Relativo}
\usage{
is_relative_path(path)
}
\arguments{
\item{path}{caminho a verificar}
}
\value{
\code{TRUE} se relativo, \code{FALSE} se absoluto ou URI
}
\description{
Verifica se Caminho e Relativo
}
\keyword{internal}
