% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.r
\name{is_valid_worker_count}
\alias{is_valid_worker_count}
\title{Verifica Validade do Numero de Workers}
\usage{
is_valid_worker_count(workers)
}
\arguments{
\item{workers}{valor a verificar}
}
\value{
\code{TRUE} se valido, \code{FALSE} caso contrario
}
\description{
Verifica Validade do Numero de Workers
}
\keyword{internal}
