% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{metrics_as_list}
\alias{metrics_as_list}
\title{Converte Metricas de Ambiente para Lista}
\usage{
metrics_as_list(metrics)
}
\arguments{
\item{metrics}{environment de metricas}
}
\value{
lista com todos os campos serializados
}
\description{
Serializa metrics environment para lista, preparado para JSON.
}
