% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{monta_dt_prev}
\alias{monta_dt_prev}
\title{Monta data.table com previsoes elaboradas para uma usina}
\usage{
monta_dt_prev(prev_usina_elem, id_usina, data_referencia)
}
\arguments{
\item{prev_usina_elem}{\code{list} contendo listas pelo menos com os elementos:
\itemize{
\item combinacao_ajuste - conjunto de parametros que definem a previsao,
incluindo id_modelo_nwp, horiz_prev, hora_min, modelo_prev
\item prev - valor previsto
}}

\item{id_usina}{\code{character} com identificador da usina}

\item{data_referencia}{\code{date} que identificada data da rodada}
}
\value{
\code{data.table} com os dados previstos para a usina
}
\description{
Organiza previsao elaborada para uma usina em um data.table
contendo informacoes do modelo de previsao, modelo nwp, data_hora_rodada
e data_hora_previsao
}
