% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{parse_predict}
\alias{parse_predict}
\title{Metodo Generico para Previsao}
\usage{
parse_predict(modelo, ...)
}
\arguments{
\item{modelo}{Objeto do modelo treinado (com classe S3 definida)}

\item{...}{Argumentos adicionais passados aos metodos especificos}
}
\value{
Vetor numerico com valores previstos
}
\description{
Funcao generica S3 que despacha para o metodo especifico de previsao
baseado na classe do modelo.
}
\seealso{
\code{\link{parse_predict.arimax}} para previsao ARIMAX
\code{\link{parse_predict.fisico_estimado}} para previsao Fisico-Estimado
}
