% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{parse_train.default}
\alias{parse_train.default}
\title{Metodo Default para parse_train}
\usage{
\method{parse_train}{default}(modelo_parametros, ...)
}
\arguments{
\item{modelo_parametros}{list, parametros do modelo}

\item{...}{Argumentos adicionais (ignorados)}
}
\value{
gera erro com tipo de modelo desconhecido
}
\description{
Metodo Default para parse_train
}
