% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{parse_train}
\alias{parse_train}
\title{Metodo Generico para Treinamento de Modelo}
\usage{
parse_train(modelo_parametros, ...)
}
\arguments{
\item{modelo_parametros}{Lista com parametros do modelo, incluindo classe S3}

\item{...}{Argumentos adicionais passados aos metodos especificos}
}
\value{
Objeto do modelo treinado (tipo depende do metodo especifico)
}
\description{
Funcao generica S3 que despacha para o metodo especifico de treinamento
baseado no tipo de modelo especificado.
}
\seealso{
\code{\link{parse_train.arimax}} para treinamento ARIMAX
\code{\link{parse_train.fisico_estimado}} para treinamento Fisico-Estimado
}
