% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_data_referencia}
\alias{parsearg_data_referencia}
\alias{parsearg_data_referencia.numeric}
\alias{parsearg_data_referencia.character}
\alias{parsearg_data_referencia.list}
\title{Interpretador De Chave \code{data_referencia}}
\usage{
parsearg_data_referencia(x)

\method{parsearg_data_referencia}{numeric}(x)

\method{parsearg_data_referencia}{character}(x)

\method{parsearg_data_referencia}{list}(x)
}
\arguments{
\item{x}{valor da chave \code{data_referencia}; numerico ou vetor de duas strings de data}
}
\value{
vetor \code{Date} de duas posicoes indicando inicio e fim da janela de simulacao
}
\description{
Funcao interna de \code{parse_config} para interpretar o parametro \code{data_referencia} da configuracao
}
