% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_horizonte_dias}
\alias{parsearg_horizonte_dias}
\title{Parser de horizonte_dias}
\usage{
parsearg_horizonte_dias(x)
}
\arguments{
\item{x}{Lista de strings com horizontes (e.g., list("D+0", "D+1"))}
}
\value{
Vetor character com horizontes
}
\description{
Converte lista de horizontes para vetor character.
}
\keyword{internal}
