% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_modelos_nwp}
\alias{parsearg_modelos_nwp}
\title{Interpretador De Chave \code{modelos_NWP}}
\usage{
parsearg_modelos_nwp(x)
}
\arguments{
\item{x}{valor da chave \code{modelos_NWP}; lista de modelos NWP}
}
\value{
vetor de strings com modelos NWP
}
\description{
Funcao interna de parse_config para interpretar o parametro \code{modelos_NWP} da configuracao
}
