% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_modelos_previsao}
\alias{parsearg_modelos_previsao}
\title{Parser de modelos_previsao}
\usage{
parsearg_modelos_previsao(x)
}
\arguments{
\item{x}{Lista com parametros do modelo, incluindo campo "tipo"}
}
\value{
Lista com classe S3 igual ao valor de "tipo"
}
\description{
Converte configuracao de modelo de previsao em objeto S3 com classe
definida pelo campo "tipo".
}
\keyword{internal}
