% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfv-sh.r
\docType{package}
\name{pfv.sh-package}
\alias{pfv.sh}
\alias{pfv.sh-package}
\title{\code{pfv.sh}}
\description{
Modelo para Previsao De Solar Fotovoltaica
}
\details{
Este pacote contem as funcionalidades necessarias para a realizacao de previsoes
de geracao solar fotovoltaica a partir de dados meteorologicos.
}
\author{
\strong{Maintainer}: Isabela Fiuza \email{isabela.fiuza@ons.org.br}

}
\keyword{internal}
