% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict_plant_wrapper}
\alias{predict_plant_wrapper}
\title{Carrega Artefato e Executa Previsao para Uma Usina}
\usage{
predict_plant_wrapper(
  iu,
  dt_usinas,
  dt_ger_obs,
  dt_prev,
  v_modelos_nwp,
  v_horizonte,
  parametros_modelo_previsao,
  parametros_periodo_geracao,
  artifact_dir,
  data_prev,
  ...
)
}
\arguments{
\item{iu}{character, identificador da usina}

\item{dt_usinas}{data.table, cadastro de usinas}

\item{dt_ger_obs}{data.table, geracao observada}

\item{dt_prev}{list, previsoes meteorologicas}

\item{v_modelos_nwp}{character, modelos NWP}

\item{v_horizonte}{character, horizontes}

\item{parametros_modelo_previsao}{list, parametros dos modelos}

\item{parametros_periodo_geracao}{list, parametros do periodo solar}

\item{artifact_dir}{character, diretorio de artefatos}

\item{data_prev}{Date, datas-alvo}

\item{...}{argumentos adicionais (ignorados)}
}
\value{
list de previsoes (retorno de \code{\link[=predict_usina]{predict_usina()}})
}
\description{
Funcao wrapper compativel com \code{run_plants()}: carrega o artefato de
modelo, valida, extrai os modelos ajustados e delega a execucao a
\code{\link[=predict_usina]{predict_usina()}}.
}
\keyword{internal}
