% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.r
\name{prov_as_list}
\alias{prov_as_list}
\title{Converte Proveniencia de Ambiente para Lista}
\usage{
prov_as_list(provenance)
}
\arguments{
\item{provenance}{environment de proveniencia}
}
\value{
lista com todos os campos, \code{plant_status} tambem como lista
}
\description{
Serializa um registro de proveniencia para lista.
}
\keyword{internal}
