% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{seleciona_modelo}
\alias{seleciona_modelo}
\title{Seleciona o melhor modelo entre ARIMA e ARIMAX}
\usage{
seleciona_modelo(nlmod1, nlmod2, erros)
}
\arguments{
\item{nlmod1}{Modelo ARIMA.}

\item{nlmod2}{Modelo ARIMAX.}

\item{erros}{Lista com erros medios dos modelos.}
}
\value{
Lista com modelo escolhido.
}
\description{
Seleciona o melhor modelo entre ARIMA e ARIMAX
}
