% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{seleciona_modelo_fisico_estimado}
\alias{seleciona_modelo_fisico_estimado}
\title{Seleciona o melhor modelo entre Regressao Linear Simples (RLS) e Regressao Linear Multipla (RLM)}
\usage{
seleciona_modelo_fisico_estimado(nlmod1, nlmod2, erros)
}
\arguments{
\item{nlmod1}{Modelo RLS}

\item{nlmod2}{Modelo RLM}

\item{erros}{Lista com erros medios dos modelos.}
}
\value{
Lista com modelo escolhido.
}
\description{
Seleciona o melhor modelo entre Regressao Linear Simples (RLS) e Regressao Linear Multipla (RLM)
}
