% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{train_modelo}
\alias{train_modelo}
\title{Treina Modelo Individual}
\usage{
train_modelo(l, ger_usi, prev_met_usi, param_modelo_previsao, data_fim_treino)
}
\arguments{
\item{l}{list com \code{id_modelo_nwp}, \code{horiz_prev}, \code{hora_min}, \code{modelo_prev}}

\item{ger_usi}{data.table, geracao observada}

\item{prev_met_usi}{list, previsoes meteorologicas}

\item{param_modelo_previsao}{list, parametros do modelo}

\item{data_fim_treino}{Date, limite para treinamento}
}
\value{
list com \code{combinacao_ajuste} e \code{modelo} ajustado
}
\description{
Treina um modelo para uma combinacao especifica.
}
\keyword{internal}
