% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valid_tipos}
\alias{valid_tipos}
\title{Valida Tipos de Uma Chave (Escalar ou Lista)}
\usage{
valid_tipos(l, tipos)
}
\arguments{
\item{l}{Valor da chave (escalar, vetor ou lista)}

\item{tipos}{Tipos esperados (character, ou lista para multiplos tipos aceitos)}
}
\value{
Logico indicando se validacao passou
}
\description{
Verifica se todos os elementos de uma chave possuem tipos validos.
}
\keyword{internal}
