% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valid_tipos_unit}
\alias{valid_tipos_unit}
\title{Valida Tipo de Um Elemento}
\usage{
valid_tipos_unit(x, tipos)
}
\arguments{
\item{x}{Elemento a validar}

\item{tipos}{Tipos esperados}
}
\value{
Logico indicando se tipo e valido
}
\description{
Funcao auxiliar para validacao unitaria de tipo.
}
\keyword{internal}
