% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valida_nomes_config}
\alias{valida_nomes_config}
\title{Valida Chaves do Arquivo de Configuracao}
\usage{
valida_nomes_config(config)
}
\arguments{
\item{config}{Lista de configuracoes}
}
\value{
\code{NULL} invisivel se valido; gera erro se chaves faltantes
}
\description{
Verifica se todas as chaves obrigatorias estao presentes.
}
\keyword{internal}
