% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{valida_tipos_config}
\alias{valida_tipos_config}
\title{Valida Tipos das Chaves de Configuracao}
\usage{
valida_tipos_config(config)
}
\arguments{
\item{config}{Lista de configuracoes}
}
\value{
\code{NULL} invisivel se valido; gera erro se tipos incorretos
}
\description{
Verifica se os valores possuem os tipos esperados.
}
\keyword{internal}
