% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{ajusta_dummy}
\alias{ajusta_dummy}
\title{Cria um modelo ARIMA dummy}
\usage{
ajusta_dummy(dt)
}
\arguments{
\item{dt}{\code{data.table} com data_hora, geracao, irradiancia,
temperatura e umidade do periodo de treinamento selecionado}
}
\value{
modelo ARIMA neutro
}
\description{
Cria um modelo ARIMA dummy
}
