% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{aplica_regressao_linear}
\alias{aplica_regressao_linear}
\title{Aplica Regressao Linear}
\usage{
aplica_regressao_linear(dt_y, dt_x, nlmod0 = NULL)
}
\arguments{
\item{dt_y}{data.table com variavel resposta}

\item{dt_x}{data.table com variaveis explicativas}

\item{nlmod0}{Modelo de fallback (retornado em caso de erro); opcional}
}
\value{
Objeto \code{lm} com o modelo ajustado
}
\description{
Ajusta modelo de regressao linear com fallback em caso de erro.
}
\keyword{internal}
