% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{aplica_regressao_linear}
\alias{aplica_regressao_linear}
\title{Aplica Regressao Linear}
\usage{
aplica_regressao_linear(dt_y, dt_x, nlmod0)
}
\arguments{
\item{dt_y}{´data.table´ contendo a variavel resposta. Exemplo: geracao observada}

\item{dt_x}{´data.table´ contendo a(s) variavel(is) explicativas. Exemplo: irradiancia, umidade e temperatura
As observacoes contidas nos data.tables contendo as variaveis ja devem estar em posicoes compativeis}

\item{nlmod0}{modelo de fallback em caso de erro no ajuste}
}
\value{
modelos ajustados
}
\description{
Funcao que aplica regressao linear a um conjunto de dados
}
