% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{associa_nwp_usina}
\alias{associa_nwp_usina}
\title{Associa Usina A Quadricula da NWP}
\usage{
associa_nwp_usina(dt_prev, dt_usinas)
}
\arguments{
\item{dt_prev}{´data.table´ contendo os dados das variaveis previstas do modelo NWP}

\item{dt_usinas}{´data.table´ contendo os dados cadastrais das usinas}
}
\value{
´data.table´ das usinas simuladas e as respectivas variaveis previstas do modelo NWP

OBSERVACAO ISABELA: Funcao aproveitada do MH,
troquei o nome do argumento dt_irrad_prev para dt_prev
para deixar generico
}
\description{
Obtem as previsoes dos modelos NWP para as usinas simuladas. Para isso, realiza a associacao das coordenadas das
usinas com quadricula correspondente do modelo NWP.
E escolhida a quadricula que apresenta menor distancia euclidiana entre seu centroide e a coordenada da usina.
}
