% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{carrega_modelo_rds}
\alias{carrega_modelo_rds}
\title{Carrega Artefato de Modelo}
\usage{
carrega_modelo_rds(modelo_previsao, id_usina, diretorio)
}
\arguments{
\item{modelo_previsao}{character, tipo do modelo}

\item{id_usina}{character, identificador da usina}

\item{diretorio}{character, caminho do diretorio}
}
\value{
Objeto do modelo carregado
}
\description{
Le arquivo RDS com modelo treinado.
}
\keyword{internal}
