% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{cria_dt_auxiliar}
\alias{cria_dt_auxiliar}
\title{Cria data.table Auxiliar para Datas Ausentes}
\usage{
cria_dt_auxiliar(dif, dt_completo)
}
\arguments{
\item{dif}{data.table com uma linha (info sobre rodada ausente)}

\item{dt_completo}{data.table de referencia (latitude/longitude)}
}
\value{
data.table com estrutura identica, valores = NA
}
\description{
Auxiliar para \code{preenche_ausencia_previsao}.
}
\keyword{internal}
