% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{cria_dt_auxiliar}
\alias{cria_dt_auxiliar}
\title{Cria data.table Auxiliar para Datas Ausentes}
\usage{
cria_dt_auxiliar(dif, dt_completo)
}
\arguments{
\item{dif}{data.table com uma linha contendo:
\describe{
\item{id_modelo_nwp}{Modelo NWP}
\item{id_usina}{Usina}
\item{data_hora_rodada}{Data da rodada ausente}
\item{passos_inicio}{Offset inicial das previsoes}
\item{passos_fim}{Offset final das previsoes}
}}

\item{dt_completo}{data.table de referencia para obter latitude/longitude}
}
\value{
data.table com estrutura identica as previsoes, valores = NA
}
\description{
Funcao auxiliar de \code{preenche_ausencia_previsao} que gera um data.table
com previsoes NA para uma data de rodada ausente.
}
\keyword{internal}
