% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{dados_suficientes}
\alias{dados_suficientes}
\title{Verifica se ha dados suficientes para o ajuste}
\usage{
dados_suficientes(dt, num_min_dados)
}
\arguments{
\item{dt}{\code{data.table} com data_hora, geracao, irradiancia,
temperatura e umidade}

\item{num_min_dados}{numero minimo de registros validos simultaneos}
}
\value{
TRUE se houver dados validos, FALSE caso contrario
}
\description{
Verifica se ha dados suficientes para o ajuste
}
\examples{
\dontrun{
dt <- data.table::data.table(
    data_hora = 1:5,
    ger = c(10, NA, 12, 0, 15),
    irr = c(200, 210, NA, 205, 215)
)
dados_suficientes(dt, num_min_dados = 2)
}
}
