% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{desnormaliza_variaveis}
\alias{desnormaliza_variaveis}
\title{Desnormaliza Variaveis}
\usage{
desnormaliza_variaveis(dt, stats_norm)
}
\arguments{
\item{dt}{data.table com colunas normalizadas (sufixo \verb{_norm})}

\item{stats_norm}{data.table com \code{variavel}, \code{med}, \code{sd}}
}
\value{
data.table desnormalizado (sem sufixo \verb{_norm})
}
\description{
Reverte normalizacao z-score, retornando valores a escala original.
}
\keyword{internal}
