% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{elimina_dados_invalidos}
\alias{elimina_dados_invalidos}
\title{Elimina Dados Invalidos}
\usage{
elimina_dados_invalidos(dt)
}
\arguments{
\item{dt}{data.table com series temporais}
}
\value{
data.table apenas com registros completos
}
\description{
Remove registros com valores 999 ou NA (conversion, then filtering).
}
\keyword{internal}
