% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{filtra_dado_por_combinacao}
\alias{filtra_dado_por_combinacao}
\title{Filtra data.table com informacao meteorologica e com geracao
para ajuste do modelo de interesse (nwp x passo de previsao x meia-hora)
e une ambos os data.tables em um unico}
\usage{
filtra_dado_por_combinacao(elem_comb, prev_met_usi, ger_usi)
}
\arguments{
\item{elem_comb}{\code{list} em que cada elemento contem:
\itemize{
\item id_modelo_nwp
\item horiz_prev
\item hora_min
}}

\item{prev_met_usi}{\code{list} contendo data.tables pelo menos com as colunas:
\itemize{
\item id_modelo_nwp - modelo meteorologico
\item data_hora_previsao - data referencia da previsao
\item horiz_prev - passo de previsao
\item hora_min - data_hora em formato HH:MM
\item valor - variavel meteorologica prevista
}}

\item{ger_usi}{\code{data.table} pelo menos com as colunas:
\itemize{
\item data_hora_observacao - data referencia da observacao
\item hora_min - data_hora em formato HH:MM
\item valor - geracao observada
}}
}
\value{
\code{data.table} com os dados filtrados para treinamento
}
\description{
Para cada trio modelo meteorologico x passo de previsao x meias-horas,
e filtrado o conjunto de dados de interesse do data.table original com
as previsoes e com a geracao. Realiza ainda a juncao em um unico data.table
com os dados para treinamento
}
