% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{get_dataset}
\alias{get_dataset}
\title{Cria Dataset Dos Dados Para Treino dos Modelos}
\usage{
get_dataset(args, conn)
}
\arguments{
\item{args}{Lista de argumentos necessarios para criacao dos datasets}

\item{conn}{Objeto de conexao com banco de dados}
}
\value{
lista contendo o dataset
}
\description{
Cria dataset dos dados usados no treinamento dos modelos de previsao de acordo configuracao especificada
}
