% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.r
\name{inner_parser_generic_args}
\alias{inner_parser_generic_args}
\title{Adiciona Argumentos Genericos ao Parser}
\usage{
inner_parser_generic_args(parser)
}
\arguments{
\item{parser}{Objeto \code{ArgumentParser}}
}
\value{
Parser com argumentos adicionados
}
\description{
Funcao interna que configura os argumentos padrao do CLI.
}
\keyword{internal}
