% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{interpola_previsao_nwp}
\alias{interpola_previsao_nwp}
\title{Interpola Dados De Previsao NWP}
\usage{
interpola_previsao_nwp(dt_prev)
}
\arguments{
\item{dt_prev}{´data.table´ com os dados de previsao NWP a serem interpolados}
}
\value{
´data.table´ com dados previstos interpolados
}
\description{
Realiza interpolação dos dados previstos dos modelos NWP em intervalos semi-horarios
}
