% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{interpola_serie_temporal}
\alias{interpola_serie_temporal}
\title{Interpola Serie Temporal}
\usage{
interpola_serie_temporal(dt)
}
\arguments{
\item{dt}{vetor numerico ou objeto a ser interpolado}
}
\value{
vetor com valores interpolados
}
\description{
Identifica as lacunas da serie temporal e preenche realizando interpolacao linear
}
