% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{normaliza_variaveis}
\alias{normaliza_variaveis}
\title{Normaliza variaveis de entrada}
\usage{
normaliza_variaveis(dt)
}
\arguments{
\item{dt}{\code{data.table} com data_hora, geracao, irradiancia,
temperatura e umidade do periodo de treinamento selecionado}
}
\value{
lista com:
\itemize{
\item \code{dados} - data.table com variaveis normalizadas
\item \code{stats} - medias e desvios padroes de cada variavel
}
}
\description{
Normaliza variaveis de entrada
}
