% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{parse_predict.default}
\alias{parse_predict.default}
\title{Metodo Default para parse_predict}
\usage{
\method{parse_predict}{default}(modelo, ...)
}
\arguments{
\item{modelo}{Objeto do modelo}

\item{...}{Argumentos adicionais (ignorados)}
}
\value{
gera erro com tipo de modelo desconhecido
}
\description{
Metodo Default para parse_predict
}
