% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{parse_predict}
\alias{parse_predict}
\title{Metodo Generico para Previsao}
\usage{
parse_predict(modelo, ...)
}
\arguments{
\item{modelo}{Objeto do modelo treinado (com classe S3)}

\item{...}{Argumentos adicionais passados aos metodos especificos}
}
\value{
Vetor numerico com valores previstos
}
\description{
Funcao generica S3 que despacha para metodo especifico de previsao.
}
