% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{parse_train.arimax}
\alias{parse_train.arimax}
\title{Treinamento Usando O Metodo ARIMAX}
\usage{
\method{parse_train}{arimax}(modelo_parametros, ...)
}
\arguments{
\item{modelo_parametros}{lista com parametros do modelo}

\item{...}{argumentos adicionais incluindo pars, ger_usi, prev_met_usi, data_fim_treino}
}
\value{
modelos ajustados
}
\description{
Treinamento Usando O Metodo ARIMAX
}
