% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{parse_train.default}
\alias{parse_train.default}
\title{Metodo Default para parse_train}
\usage{
\method{parse_train}{default}(modelo_parametros, ...)
}
\arguments{
\item{modelo_parametros}{Lista com parametros do modelo}

\item{...}{Argumentos adicionais (ignorados)}
}
\value{
Gera erro indicando tipo de modelo desconhecido
}
\description{
Metodo default que gera erro para tipos de modelo nao suportados.
}
