% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{parse_train.fisico_estimado}
\alias{parse_train.fisico_estimado}
\title{Treinamento Usando O Metodo Fisico Estimado}
\usage{
\method{parse_train}{fisico_estimado}(modelo_parametros, ...)
}
\arguments{
\item{modelo_parametros}{lista com parametros do modelo}

\item{...}{argumentos adicionais incluindo pars, ger_usi, prev_met_usi, data_fim_treino}
}
\value{
modelos ajustados
}
\description{
Realiza treinamento do modelo fisico estimado e salva modelo para uso futuro
}
