% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{parse_train}
\alias{parse_train}
\title{Metodo Generico para Treinamento de Modelo}
\usage{
parse_train(modelo_parametros, ...)
}
\arguments{
\item{modelo_parametros}{list, parametros do modelo}

\item{...}{Argumentos adicionais passados aos metodos especificos}
}
\value{
Objeto do modelo treinado
}
\description{
Funcao generica S3 que despacha para metodo especifico de treinamento.
}
