% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_data_referencia}
\alias{parsearg_data_referencia}
\alias{parsearg_data_referencia.numeric}
\alias{parsearg_data_referencia.character}
\alias{parsearg_data_referencia.list}
\title{Parser de data_referencia}
\usage{
parsearg_data_referencia(x)

\method{parsearg_data_referencia}{numeric}(x)

\method{parsearg_data_referencia}{character}(x)

\method{parsearg_data_referencia}{list}(x)
}
\arguments{
\item{x}{Valor da chave data_referencia}
}
\value{
Vetor Date ou lista de Dates
}
\description{
Interpreta o parametro de data de referencia, que pode ser:
\itemize{
\item Numerico: numero de dias antes de hoje
\item Character: data no formato "YYYY-MM-DD"
\item Lista: combinacao dos anteriores
}
}
\details{
Para input numerico N, retorna intervalo de N+1 dias antes de hoje

Para input character, converte diretamente para Date

Para input lista, processa cada elemento recursivamente
}
