% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_horizonte_dias}
\alias{parsearg_horizonte_dias}
\title{Interpretador De Chave \code{horizonte_dias}}
\usage{
parsearg_horizonte_dias(x)
}
\arguments{
\item{x}{valor da chave \code{horizonte_dias}; lista de strings com horizontes}
}
\value{
vetor de strings com horizontes
}
\description{
Funcao interna de parse_config para interpretar o parametro \code{horizonte_dias} da configuracao
}
