% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_ids_usinas}
\alias{parsearg_ids_usinas}
\title{Parser de ids_usinas}
\usage{
parsearg_ids_usinas(x, conn)
}
\arguments{
\item{x}{Lista de IDs de usinas (pode ser vazia)}

\item{conn}{Objeto de conexao com banco de dados}
}
\value{
Vetor character com IDs unicos de usinas
}
\description{
Interpreta lista de IDs de usinas. Se vazia, retorna todas as usinas
disponiveis no banco de dados.
}
\keyword{internal}
