% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_modelos_nwp}
\alias{parsearg_modelos_nwp}
\title{Parser de modelos_NWP}
\usage{
parsearg_modelos_nwp(x)
}
\arguments{
\item{x}{Lista de strings com modelos NWP (e.g., list("GFS", "ECMWF"))}
}
\value{
Vetor character com modelos NWP
}
\description{
Converte lista de modelos NWP para vetor character.
}
\keyword{internal}
