% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-file.r
\name{parsearg_modelos_previsao}
\alias{parsearg_modelos_previsao}
\title{Interpretador De Chave \code{modelos_previsao}}
\usage{
parsearg_modelos_previsao(x)
}
\arguments{
\item{x}{valor da chave \code{modelos_previsao}; lista com parametros do modelo}
}
\value{
lista com classe definida pelo tipo do modelo
}
\description{
Funcao interna de parse_config para interpretar o parametro \code{modelos_previsao} da configuracao
}
