% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict_usina}
\alias{predict_usina}
\title{Gera Previsoes para Uma Usina}
\usage{
predict_usina(
  iu,
  dt_usinas,
  dt_ger_obs,
  dt_prev,
  v_modelos_nwp,
  v_horizonte,
  parametros_modelo_previsao,
  parametros_periodo_geracao,
  models,
  data_prev
)
}
\arguments{
\item{iu}{character, identificador da usina}

\item{dt_usinas}{data.table, cadastro de usinas}

\item{dt_ger_obs}{data.table, geracao observada}

\item{dt_prev}{list, previsoes meteorologicas}

\item{v_modelos_nwp}{character, modelos NWP}

\item{v_horizonte}{character, horizontes}

\item{parametros_modelo_previsao}{list, parametros dos modelos}

\item{parametros_periodo_geracao}{list, parametros do periodo solar}

\item{models}{list, modelos ajustados}

\item{data_prev}{Date, datas-alvo}
}
\value{
list de previsoes (uma por combinacao treinada)
}
\description{
Executa previsao para uma usina especifica.
}
\keyword{internal}
