% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{preenche_ausencia_previsao}
\alias{preenche_ausencia_previsao}
\title{Preenche Lacunas de Rodadas NWP}
\usage{
preenche_ausencia_previsao(dt)
}
\arguments{
\item{dt}{data.table com previsoes meteorologicas}
}
\value{
data.table com datas de rodada completas
}
\description{
Identifica datas ausentes de rodadas NWP e preenche com NA.
}
\keyword{internal}
