% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{preenche_ausencia_previsao}
\alias{preenche_ausencia_previsao}
\title{Preenche Lacunas de Previsao}
\usage{
preenche_ausencia_previsao(dt)
}
\arguments{
\item{dt}{'data.table' contendo as previsoes meteorologicas}
}
\value{
'data.table' contendo as previsoes com datas de execucao do modelo meteorologico completas
}
\description{
Identifica dias em que nao ha rodada do modelo meteorologico na base de dados e preenche as previsoes
referentes a essa execucao ausente com NA.
A funcao avalia a coluna 'data_hora_rodada' de cada modelo meteorologico e, sendo verificada a
ausencia de alguma data, inclui dados NA para compatibilizacao das series temporais.
}
