% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{recalibra_arima}
\alias{recalibra_arima}
\title{Recalibra Modelo ARIMA}
\usage{
recalibra_arima(dt, nlmod0)
}
\arguments{
\item{dt}{data.table com coluna \code{ger_obs_norm} (geracao normalizada)}

\item{nlmod0}{Modelo ARIMA original}
}
\value{
Objeto Arima recalibrado
}
\description{
Reajusta ARIMA mantendo estrutura (p,d,q) com dados atualizados.
}
\keyword{internal}
