% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{recalibra_arima}
\alias{recalibra_arima}
\title{Recalibra modelo ARIMA simples}
\usage{
recalibra_arima(dt, nlmod0)
}
\arguments{
\item{dt}{data.table com variavel \code{ger_obs_norm}.}

\item{nlmod0}{modelo ajustado pela etapa de treinamento.}
}
\value{
modelo recalibrado do tipo \code{Arima}.
}
\description{
Recalibra modelo ARIMA simples
}
