% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{recalibra_arimax}
\alias{recalibra_arimax}
\title{Recalibra modelo ARIMAX com variaveis exogenas}
\usage{
recalibra_arimax(dt, nlmod0)
}
\arguments{
\item{dt}{data.table com variavel \code{ger_obs_norm}, \code{irrad_prev_norm},
\code{temp_prev_norm} e \code{umid_prev_norm}.}

\item{nlmod0}{modelo ajustado pela etapa de treinamento.}
}
\value{
modelo recalibrado do tipo \code{Arimax}.
}
\description{
Recalibra modelo ARIMAX com variaveis exogenas
}
