% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{recalibra_arimax}
\alias{recalibra_arimax}
\title{Recalibra Modelo ARIMAX}
\usage{
recalibra_arimax(dt, nlmod0)
}
\arguments{
\item{dt}{data.table com \code{ger_obs_norm} e variaveis exogenas normalizadas}

\item{nlmod0}{Modelo ARIMAX original}
}
\value{
Objeto Arima recalibrado com variaveis exogenas
}
\description{
Reajusta ARIMAX mantendo estrutura com dados atualizados.
}
\keyword{internal}
