% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{seleciona_janela}
\alias{seleciona_janela}
\title{Seleciona janela dos dados para treinamento}
\usage{
seleciona_janela(dt, data_ref, janela_dias_treinamento)
}
\arguments{
\item{dt}{\code{data.table} com data_hora, geracao, variaveis meteorologicas}

\item{data_ref}{data de referencia para selecao da janela}

\item{janela_dias_treinamento}{numero de dias utilizados
para treinamento}
}
\value{
subset do data.table filtrado
}
\description{
Seleciona janela dos dados para treinamento
}
