% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.r
\name{train_modelo}
\alias{train_modelo}
\title{Treina Modelo Individual}
\usage{
train_modelo(l, ger_usi, prev_met_usi, param_modelo_previsao, data_fim_treino)
}
\arguments{
\item{l}{Lista contendo a combinacao de parametros:
\describe{
\item{id_modelo_nwp}{Identificador do modelo NWP}
\item{horiz_prev}{Horizonte de previsao (e.g., "D+0")}
\item{hora_min}{Meia-hora no formato "HH:MM"}
\item{modelo_prev}{Tipo de modelo de previsao}
}}

\item{ger_usi}{data.table com geracao observada da usina}

\item{prev_met_usi}{Lista de data.tables com previsoes meteorologicas}

\item{param_modelo_previsao}{Lista com parametros dos modelos}

\item{data_fim_treino}{Data limite para dados de treinamento}
}
\value{
Lista contendo:
\describe{
\item{combinacao_ajuste}{Parametros da combinacao}
\item{modelo}{Objeto do modelo ajustado}
}
}
\description{
Treina um modelo de previsao para uma combinacao especifica de
NWP, horizonte, meia-hora e tipo de modelo.
}
\keyword{internal}
